requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-storages==1.12.3
gcloud==0.18.3
google-api-core==2.2.1
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-automl==2.5.1
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.1
hurry.filesize==0.9
idna==3.3
jws==0.1.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.2
Pillow==8.4.0
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
PyJWT==2.3.0
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
